refactor(generator): unfromat `components.book.dart` (#1366)
The `$format` function didn't work well when we allowed [`source_gen`
v2.0.0](https://pub.dev/packages/source_gen/changelog#200), because it
has a Breaking Change of adding new parameter. Thus, we drop the
formatting as a whole for this file, since it's a part of the SAM
architecture, and doesn't have a priority in v3.