refactor: adjust mobile breakpoint (#1416)
The desktop layout does not really work on a tablet in portrait
orientation. The navigation has many overflows and is generally too
small to actually read any use case names, etc. The same applies to the
knobs and addons UI on the right.
According to https://m3.material.io/foundations/layout/applying-layout/window-size-classes
the limit for desktop, landscape tablet, etc. is 840, not 800. With 840,
our tablets also now correctly switch between desktop for landscape and
mobile for portrait.