feat: add `home` widget (#1418)
The home widget is a widget that is shown on startup when no use-case is
selected. This widget does not inherit from the `appBuilder` or the
`addons`; Meaning that if `Theme.of(context)` is called inside this
widget, then it will use Widgetbook's `lightTheme` or `darkTheme`, and
not the Theme from the `appBuilder` or the `ThemeAddon`.

![image](https://github.com/user-attachments/assets/9c66e7af-38af-4ed7-8ab6-c04e7d0952bc)