refactor: rework nullable knobs (#1450)
Previously, the null state of the knob was held inside the `Knob` class,
and not reflected in the URL. This prevented the `nullable` knobs from
being supported in the `KnobConfig` for multi-snapshot feature.

With this PR, nullable knobs can be expressed as follows:
1. `/?path=my-use-case&knobs={my-knob:??}` - just null
2. `/?path=my-use-case&knobs={my-knob:??my-value}` - `my-value` will be
used when the knob is activated via the checkbox.