deps: bump sdk to 3.7 & flutter to 3.29 (#1541)
We are taking a leap from Flutter 3.19 to 3.29 as the minimum required
version for Widgetbook.
We would have loved to support a wider range of Flutter versions, but we
are forced to update our minimum SDK to 3.7 (i.e. Flutter 3.29) to be
able to support build v3 (#1528) and the new analyzer APIs.

> [!NOTE]
> This PR contains a lot of formatting changes that were introduced in
dart 3.7's tall-style